requirements.txt
certifi==2020.12.5
chardet==4.0.0
coloredlogs==14.0
greenlet==1.0a1
humanfriendly==9.0
idna==2.10
playwright==0.162.2
prettytable==2.0.0
pyee==8.1.0
pyreadline==2.1
PyYAML==5.3.1
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
